refactor(dao): dao 인터페이스화
