+++ src/test/java/racingcar/dao/DbCarDaoTest.java
@@ -18,9 +18,9 @@ import racingcar.dto.RacingCarResultDto;
 
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
 @JdbcTest
-class CarDaoTest {
-    private CarDao carDao;
-    private GameDao gameDao;
+class DbCarDaoTest {
+    private DbCarDao dbCarDao;
+    private DbGameDao dbGameDao;
 
     private List<RacingCarResultDto> cars;
 
@@ -29,11 +29,11 @@ class CarDaoTest {
 
     @BeforeEach
     void setUp() {
-        carDao = new CarDao(jdbcTemplate);
-        gameDao = new GameDao(jdbcTemplate.getJdbcTemplate().getDataSource());
+        dbCarDao = new DbCarDao(jdbcTemplate);
+        dbGameDao = new DbGameDao(jdbcTemplate.getJdbcTemplate().getDataSource());
         jdbcTemplate.getJdbcTemplate().execute("ALTER TABLE game ALTER COLUMN id RESTART WITH 1");
-        long id1 = gameDao.save(1);
-        long id2 = gameDao.save(2);
+        long id1 = dbGameDao.save(1);
+        long id2 = dbGameDao.save(2);
         cars = List.of(
                 RacingCarResultDto.of(new RacingCar("오잉"), GameResult.LOSE.getValue(), id1),
                 RacingCarResultDto.of(new RacingCar("포이"), GameResult.WIN.getValue(), id1),
@@ -44,8 +44,8 @@ class CarDaoTest {
     void 자동차_저장_테스트() {
         //given
         //when
-        carDao.saveAll(cars);
-        List<RacingCarDto> queriedCars = carDao.findCarsById(1);
+        dbCarDao.saveAll(cars);
+        List<RacingCarDto> queriedCars = dbCarDao.findCarsById(1);
 
         //then
         assertThat(queriedCars).hasSize(2)
